fix: proper support for C2022 in the frontend
