+++ frontend/src/pages/planner/utils/planBoardFunctions.tsx
@@ -1,5 +1,5 @@
 import { type CourseId, type CoursePos, type PseudoCourseId } from './Types'
-import { type ValidatablePlan } from '../../../client'
+import { type Cyear, type ValidatablePlan } from '../../../client'
 
 export const getCoursePos = (prevCourses: PseudoCourseId[][], courseId: CourseId): CoursePos | null => {
   const positions: CoursePos[] = []
@@ -64,4 +64,23 @@ export const updateClassesState = (prev: ValidatablePlan, drag: CoursePos, drop:
   }
 
   return { ...prev, classes: newClasses }
+}
+
+// Ensure that an array stays in sync with a union of string literals
+// https://stackoverflow.com/a/70694878/5884836
+type ValueOf<T> = T[keyof T]
+type NonEmptyArray<T> = [T, ...T[]]
+type MustInclude<T, U extends T[]> = [T] extends [ValueOf<U>] ? U : never
+function stringUnionToArray<T> () {
+  return <U extends NonEmptyArray<T>>(...elements: MustInclude<T, U>) => elements
+}
+export const VALID_CYEARS = stringUnionToArray<Cyear['raw']>()('C2020', 'C2022')
+
+export const validateCyear = (raw: string): Cyear | null => {
+  for (const cyear of VALID_CYEARS) {
+    if (raw === cyear) {
+      return { raw: cyear }
+    }
+  }
+  return null
 }
